Monkeypatch res.writeHead instead of res.end
The middleware would fail if res.write() or res.writeHead() are called, with the error:

Error [ERR_HTTP_HEADERS_SENT]: Cannot set headers after they are sent to the client

This change moves the logic to the last valid moment to make header changes.
However, note that the writeHead method takes an optional `headers` argument,
that this code will not fix up. If you want to address that edge case, you could add
apply headers argument to the request first, then don't send the headers through
to writeHead().